Ref: SIgn Up, In 코드 리팩토링
* SIgn Up, In 코드 리팩토링
